Enable wildcard with prefix.
Use case:

We use the Aurora scheduler in Mesos which assign tasks names using a
jobkey which consists of: "role/environment/jobname".  We often end up
in a scenario where we want all jobs running under a given "role" or
"role/environment" combination to have access to a Vault role.  Under
the current implementation this would require us to iterate all of the
possible combination and install policy files.

Previous versions of Vault gatekeeper allowed us to do wildcards with
prefixes as they used a simple glob match.  Moving to the radix tree
implementation broke this behaviour.

This patch changes the matching behaviour slightly.  Whenever we detect
a wildcard match we flag the policy as such.

When performing policy matching for wildcards we do a prefix search,
rather than looking for the last character being ":". Policy matching
for non wildcard policies is unchanged.

No changes to the radix tree structure are required.